Fix test names and refactor GitHubSvg's tooltip
