+++ components/SvgDisplayer/GitHubSvg/GitHubSvg.test.js
@@ -59,133 +59,6 @@ describe('<GitHubSvg />', () => {
     expect(gitHubSvgWrapper.find(GitHubHeader)).toHaveLength(1);
   });
 
-  describe('showToolTip', () => {
-    const dataCount = '2';
-    const dataDate = '2019-01-01';
-    let event;
-    let toolTipDiv;
-    let target;
-    let getElementByIdSpy;
-
-    beforeEach(() => {
-      toolTipDiv = document.createElement('div');
-      target = document.createElement('rect');
-      target.setAttribute('data-count', dataCount);
-      target.setAttribute('data-date', dataDate);
-
-      event = {
-        target,
-      };
-    });
-
-    it('appends the text node for the tooltip\'s div', () => {
-      const text = `${dataCount} contributions on ${dataDate}`;
-      const appendChildSpy = jest.fn();
-      toolTipDiv.appendChild = appendChildSpy;
-      getElementByIdSpy = jest.fn().mockReturnValue(toolTipDiv);
-      global.document.getElementById = getElementByIdSpy;
-
-      gitHubSvgWrapper.instance().constructor.showToolTip(event);
-
-      expect(appendChildSpy).toHaveBeenCalledWith(document.createTextNode(text));
-    });
-  });
-
-  describe('hideToolTip', () => {
-    let toolTipDiv;
-
-    beforeEach(() => {
-      toolTipDiv = document.createElement('div');
-    });
-
-    describe('when the toolTip doesn\'t have child', () => {
-      let removeChildSpy;
-
-      beforeEach(() => {
-        removeChildSpy = jest.fn();
-        toolTipDiv.removeChild = removeChildSpy;
-        global.document.getElementById = jest.fn().mockReturnValue(toolTipDiv);
-      });
-
-      it('doesn\'t removes the child', () => {
-        gitHubSvgWrapper.instance().constructor.hideToolTip();
-
-        expect(removeChildSpy).not.toHaveBeenCalled();
-      });
-    });
-
-    describe('when the toolTip has child', () => {
-      let toolTipChild;
-      let removeChildSpy;
-
-      beforeEach(() => {
-        toolTipChild = document.createTextNode('');
-        toolTipDiv.appendChild(toolTipChild);
-        removeChildSpy = jest.fn();
-        toolTipDiv.removeChild = removeChildSpy;
-        global.document.getElementById = jest.fn().mockReturnValue(toolTipDiv);
-      });
-
-      it('removes the child', () => {
-        gitHubSvgWrapper.instance().constructor.hideToolTip();
-
-        expect(removeChildSpy).toHaveBeenCalledWith(toolTipChild);
-      });
-    });
-  });
-
-  describe('addToolTipEventListeners', () => {
-    let rectElements;
-    let getElementsByTagNameSpy;
-    let addEventListenerSpy;
-
-    beforeEach(() => {
-      rectElements = [document.createElement('rect')];
-      addEventListenerSpy = jest.fn();
-      rectElements[0].addEventListener = addEventListenerSpy;
-      getElementsByTagNameSpy = jest.fn().mockReturnValue(rectElements);
-      global.document.getElementsByTagName = getElementsByTagNameSpy;
-    });
-
-    it('adds "mouseover" eventListener with GitHubSvg.showToolTip function', () => {
-      gitHubSvgWrapper.instance().constructor.addToolTipEventListeners();
-
-      expect(addEventListenerSpy).toHaveBeenCalledWith('mouseover', gitHubSvgWrapper.instance().constructor.showToolTip);
-    });
-
-    it('adds "mouseleave" eventListener with GitHubSvg.hideToolTip function', () => {
-      gitHubSvgWrapper.instance().constructor.addToolTipEventListeners();
-
-      expect(addEventListenerSpy).toHaveBeenCalledWith('mouseleave', gitHubSvgWrapper.instance().constructor.hideToolTip);
-    });
-  });
-
-  describe('removeToolTipEventListeners', () => {
-    let rectElements;
-    let getElementsByTagNameSpy;
-    let removeEventListenerSpy;
-
-    beforeEach(() => {
-      rectElements = [document.createElement('rect')];
-      removeEventListenerSpy = jest.fn();
-      rectElements[0].removeEventListener = removeEventListenerSpy;
-      getElementsByTagNameSpy = jest.fn().mockReturnValue(rectElements);
-      global.document.getElementsByTagName = getElementsByTagNameSpy;
-    });
-
-    it('removes "mouseover" eventListener with GitHubSvg.showToolTip function', () => {
-      gitHubSvgWrapper.instance().constructor.removeToolTipEventListener();
-
-      expect(removeEventListenerSpy).toHaveBeenCalledWith('mouseover', gitHubSvgWrapper.instance().constructor.showToolTip);
-    });
-
-    it('removes "mouseleave" eventListener with GitHubSvg.hideToolTip function', () => {
-      gitHubSvgWrapper.instance().constructor.removeToolTipEventListener();
-
-      expect(removeEventListenerSpy).toHaveBeenCalledWith('mouseleave', gitHubSvgWrapper.instance().constructor.hideToolTip);
-    });
-  });
-
   describe('fetchFirstGitHubUserCalendar', () => {
     it('sets `isLoading` to false', async () => {
       await gitHubSvgWrapper.instance().fetchFirstGitHubUserCalendar();
@@ -417,18 +290,6 @@ describe('<GitHubSvg />', () => {
       gitHubSvgWrapper.instance().writeState(data);
 
       expect(gitHubSvgWrapper.state('actualCalendar')).toEqual(data.updatedActualCalendar);
-    });
-  });
-
-  describe('componentWillUnmount', () => {
-    const removeEventListenerSpy = jest.fn();
-
-    it('calls removeToolTipEventListener function', () => {
-      gitHubSvgWrapper.instance().constructor.removeToolTipEventListener = removeEventListenerSpy;
-
-      gitHubSvgWrapper.instance().componentWillUnmount();
-
-      expect(removeEventListenerSpy).toHaveBeenCalled();
     });
   });
 });
